Misc clean up

+++ openstudiocore/src/utilities/geometry/FloorplanJS.cpp
@@ -544,7 +544,7 @@ namespace openstudio{
     double tol = 0.001;
     faceVertices = simplify(faceVertices, false, tol);
 
-    unsigned numPoints = faceVertices.size();
+    size_t numPoints = faceVertices.size();
     if (numPoints < 3){
       //LOG(Error, "Cannot create a space for floorPrint of size " << faceVertices.size() << ".");
       return;
@@ -608,7 +608,7 @@ namespace openstudio{
       Point3dVectorVector allFinalShadeVertices;
       std::vector<unsigned> allFinalShadeParentSubSurfaceIndices;
 
-      unsigned windowN = windowCenterVertices.size();
+      size_t windowN = windowCenterVertices.size();
       OS_ASSERT(windowN == windowDefinitionIds.size());
       for (unsigned windowIdx = 0; windowIdx < windowN; ++windowIdx){
         if (mappedWindows.find(windowIdx) == mappedWindows.end()){
@@ -655,7 +655,7 @@ namespace openstudio{
                 windowVertices.push_back(Point3d(window2.x(), window2.y(), sillHeight));
                 windowVertices.push_back(Point3d(window2.x(), window2.y(), sillHeight + height));
 
-                unsigned parentSubSurfaceIndex = allFinalWindowVertices.size();
+                size_t parentSubSurfaceIndex = allFinalWindowVertices.size();
                 allFinalWindowVertices.push_back(windowVertices);
 
                 if (checkKeyAndType(*windowDefinition, "overhang_projection_factor", Json::realValue) || checkKeyAndType(*windowDefinition, "overhang_projection_factor", Json::intValue)){
@@ -732,7 +732,7 @@ namespace openstudio{
                 if (test)
                 {
                   if (!viewVertices.empty()){
-                    unsigned parentSubSurfaceIndex = allFinalWindowVertices.size();
+                    size_t parentSubSurfaceIndex = allFinalWindowVertices.size();
                     allFinalWindowVertices.push_back(viewVertices);
 
                     if (!exteriorShadingVertices.empty()){
